System.exitOnNoWindowStartup を追加。 false にすると、startup.tjs 終了後にウインドウが生成されていなくても終了しなくなります。 squirrel など外部の言語系に完全に終了処理を譲る場合に設定します。
git-svn-id: svn://localhost/kirikiri2/trunk@4578 131d44d2-bf85-024a-9602-49a0ea93a7a9

+++ kirikiri2/src/core/base/win32/ScriptMgnImpl.cpp
@@ -313,7 +313,7 @@ void TVPInitializeStartupScript()
 	TVPStartObjectHashMap();
 
 	TVPExecuteStartupScript();
-	if(TVPGetWindowCount() == 0 && (TVPMainForm && !TVPMainForm->Visible))
+	if(TVPTerminateOnNoWindowStartup && TVPGetWindowCount() == 0 && (TVPMainForm && !TVPMainForm->Visible))
 	{
 		// no window is created and main window is invisible
 		Application->Terminate();
